Updated camera, added damage, some minor issuse to fix
-removed camera function cameraCollision (now calculus performed
directly oin the if statement)
-removed some other unnecessary variables
-added damage
-fixed camera behaviour, camera is now complete

TO DO:
- fix some collision toggles collidedLeftX should sometimes be
collidedRightX and collidedLeftZ should be collidedRightZ or viceversa
when wheel is going in opposite direction (problems on corner sections,
not on straight lines). In these sections camera does not update
accordingly because of the wrong toggle.
Damage now affects performance, now set to a very low value, but with a
powerup (sensible) we might change it during the race too